Moved test schema outside of dev schema
